+++ package.json
@@ -8,7 +8,8 @@
     "example": "example"
   },
   "workspaces": [
-    "packages/*"
+    "packages/*",
+    "packages/@almin/*"
   ],
   "license": "MIT",
   "devDependencies": {
