+++ packages/@almin/usecase-bus/src/index.ts
@@ -1,0 +1,1 @@
+export { UseCaseBus } from "./UseCaseBus";
