--- packages/@almin/usecase-container/src/index.ts
@@ -1,1 +1,0 @@
-export { UseCaseContainer } from "./UseCaseContainer";
